Separate elements page from container and content component
